refactor(usage): give the page's copy a module of its own
Strings lived in three places. overLimit mixed computing an overage with
writing prose about it, utils carried the plan section and the project
contribution, and the two banners said the same thing in two slightly
different sentences as a result.

copy.ts now holds everything the page says about a plan and its limit,
so someone asking what a customer is told reads one file. overLimit
keeps the arithmetic, utils keeps periods and windows, and the tests
split the same way: copy.test asserts prose, overLimit.test asserts
numbers.
